Retrieve introspection information from filesystem by default
http-dbus-bridge needs to know what arguments any method it calls has.
It can query the object with a DBus Introspect call to find this
information.  This requires a round-trip across DBus to accomplish.

This commit changes the behaviour such that it will first look on the
filesystem for a file named interface-XXX.xml where XXX is the name of
the interface to be called.  This will provide us with greater
flexibility in the future as extra annotations can be added to this
introspection xml.  As a side effect it lessens the massive security
hole as only known interfaces will now be able to be called.

To return to the old behaviour pass the option `--allow-introspection`.