Add builtin support for Exposing an object's properties over HTTP
The previous method of adding the GetAll, Get and Set methods to the
configuration file, although elegant, was not using the right types
when calling `Set` as Set takes a DBus Variant and you need to look
in the introspection XML to work out what the real type should be.

This is still not perfect as we only support simple DBus types.
More work will be required to support compound types.  The addition
of the explicit HTTP-DBus path mapping file will also be useful for
implementing transforming DBus Object Path types when converting to/
from JSON, although this work is not yet done.